refactor: userStore -> userInfoStore
